requirements.txt
streamlit
python-dotenv
easyocr
spacy
opencv-python-headless
pdf2image
pandas
openpyxl
Pillow
numpy
python-docx
PyMuPDF
openai
